+++ AlchemistCode/Cards/Common/Rectify.cs
@@ -1,0 +1,27 @@
+using System.Linq;
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.CardSelection;
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+
+namespace Alchemist.AlchemistCode.Cards.Common;
+
+public class Rectify : AlchemistCard
+{
+    public Rectify() : base(1, CardType.Skill, CardRarity.Common, TargetType.Self)
+    {
+        WithCards(2, 1);
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await CommonActions.Draw(this, choiceContext);
+        var picked = (await CardSelectCmd.FromHandForDiscard(
+            choiceContext, Owner,
+            new CardSelectorPrefs(CardSelectorPrefs.DiscardSelectionPrompt, 1),
+            null, this)).FirstOrDefault();
+        if (picked != null)
+            await CardCmd.Discard(choiceContext, picked);
+    }
+}