+++ AlchemistCode/Cards/Common/Taint.cs
@@ -8,9 +8,9 @@ namespace Alchemist.AlchemistCode.Cards.Common;
 
 public class Taint : AlchemistCard
 {
-    public Taint() : base(1, CardType.Skill, CardRarity.Common, TargetType.AnyEnemy)
+    public Taint() : base(0, CardType.Skill, CardRarity.Common, TargetType.AnyEnemy)
     {
-        WithPower<PoisonPower>(4, 2);
+        WithPower<PoisonPower>(3, 2);
         WithVar("SelfPoison", 1, 0);
     }