+++ AlchemistCode/Cards/Uncommon/Warded.cs
@@ -6,18 +6,18 @@ using MegaCrit.Sts2.Core.HoverTips;
 
 namespace Alchemist.AlchemistCode.Cards.Uncommon;
 
-public class Slag : AlchemistCard
+public class Warded : AlchemistCard
 {
-    public Slag() : base(1, CardType.Power, CardRarity.Uncommon, TargetType.Self)
+    public Warded() : base(1, CardType.Power, CardRarity.Uncommon, TargetType.Self)
     {
-        WithVar("Block", 2, 1);
+        WithVar("Block", 3, 1);
         WithTip(StaticHoverTip.Block);
-        WithTip(CardKeyword.Exhaust);
+        WithTip(typeof(AntitoxinPower));
     }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
-        await PowerCmd.Apply<SlagPower>(choiceContext, Owner.Creature,
+        await PowerCmd.Apply<WardedPower>(choiceContext, Owner.Creature,
             DynamicVars["Block"].IntValue, Owner.Creature, this);
     }
 }